+++ android/versioned-abis/expoview-abi49_0_0/src/main/java/abi49_0_0/expo/modules/updates/UpdatesModule.kt
@@ -52,7 +52,6 @@ class UpdatesModule : Module() {
 
         constants["isEmergencyLaunch"] = updatesController.isEmergencyLaunch
         constants["isEmbeddedLaunch"] = isEmbeddedLaunch
-        constants["isMissingRuntimeVersion"] = configuration.isMissingRuntimeVersion
         constants["isEnabled"] = configuration.isEnabled
         constants["releaseChannel"] = configuration.releaseChannel
         constants["isUsingEmbeddedAssets"] = updatesController.isUsingEmbeddedAssets
@@ -79,14 +78,6 @@ class UpdatesModule : Module() {
       } catch (e: Exception) {
         // do nothing; this is expected in a development client
         constants["isEnabled"] = false
-
-        // In a development client, we normally don't have access to the updates configuration, but
-        // we should attempt to see if the runtime/sdk versions are defined in AndroidManifest.xml
-        // and warn the developer if not. This does not take into account any extra configuration
-        // provided at runtime in MainApplication.java, because we don't have access to that in a
-        // debug build.
-        val configuration = UpdatesConfiguration(context, null)
-        constants["isMissingRuntimeVersion"] = configuration.isMissingRuntimeVersion
       }
       constants
     }